add sms parsing, manual input, error modals
