+++ src/components/Login/Password.js
@@ -231,8 +231,8 @@ export class Password extends Component {
     // TODO: set real ID;
     const isEmulator = DeviceInfo.isEmulator();
     const randomImei = Math.floor((10000000 + Math.random()) * 90000000).toString();
-    const imei = isEmulator ? randomImei : IMEI.getImei();
-    // const imei = randomImei.toString();
+    // const imei = isEmulator ? randomImei : IMEI.getImei();
+    const imei = randomImei.toString();
     // const imei = '1111111111925';
 
     this.props.signup({
